#338 Control publish/unpublish with roles
Add a dmc-admin and dmc-author role with proper privileges.  Use these
roles to control access to publishing and unpublishing content.  Add
new authorize module and function authorize:is-admin() to perform the
check of access level.  Add check to publish-unpublish-doc.xqy and also
show/hide link in UI based on role.