#378 add handling for api:usage attribute `camel-case`.
While implementing this I noticed that the assumptions made
in `stp:schema-info` don't really hold up. That function now logs
warnings when it can't find the expected schema info.

These messages should include enough information
to track down the problems. For example:

```
[apidoc/setup/stp:schema-info] No xs:element found for cts:base-reference in fn:doc("/8.0/apidoc/manage-database-properties.xml")/xs:schema/xs:complexType[1]/xs:sequence/xs:element source fn:doc("/8.0/apidoc/management.xml")/apidoc:module/apidoc:function[33]/apidoc:usage
[apidoc/setup/stp:schema-info] No xs:documentation found for meta in fn:doc("/8.0/apidoc/meters.xml")/xs:schema/xs:complexType[4]/xs:sequence/xs:element[1] source fn:doc("/8.0/apidoc/management.xml")/apidoc:module/apidoc:function[34]/apidoc:usage
```

The first message means that `cts:base-reference` wasn't found
for `management.xml`, because that element defined in the cts schema
and `stp:schema-info` expects all the documentation for a function
to live in one schema.

The second message means that `meta` has a schema element
for `management.xml`, but it doesn't have a descendent element
matching `xs:documentation/xs:annotation`.