#466 Test cases for REST and non-REST href fragment forms, with code changes so they pass. This includes REST hrefs such as `#POST:/manage/v2/databases/{id|name}#ValidateBackup`.
Legacy change: the code for finding target functions seemed to be broken
for non-REST functions too. It was looking for `apidoc:function/@id`,
which does not exist in the raw content I have. I changed `@id` to `@name`:
I'm not sure how that matches up with existing practice,
but at least the attribute exists.