Update homepage hero (#712)
* changing out hero images

changing out hero images to use newer watercolor branding

* adding css to support new imagery

adding css to support new imagery.

The media query code is to support when the headline goes to 4 lines in
mobile.  This is not an ideal solution but should work for short term.
Longer term we should consider CSS middle alignment and background
responsive swapping like the data interchange feature in zurb
foundation.